backend/requirements.txt
# Django Core
autoflake==2.3.1

# Django Core
python-dotenv==1.0.1

# Celery
celery==5.3.1

Django==5.0.6
djangorestframework==3.15.1

# Editor
django-ckeditor-5==0.2.18

# File Upload
django-cleanup==8.1.0

# Authentication & Security
django-cors-headers==4.3.1
django-filter==24.2

# Database
psycopg2-binary==2.9.9  # PostgreSQL

# Image Processing
Pillow==10.3.0

# Slug & UUID
python-slugify==8.0.4

# Documentation
drf-spectacular==0.27.2
django-extensions==3.2.3

# Testing
ipython==8.37.0

# Formatting
isort==8.0.1

# Development
black==24.4.2
flake8==7.0.0
pytest==8.2.0
pytest-django==4.8.0

# Production
gunicorn==22.0.0
whitenoise==6.6.0

# Redis (for caching/sessions)
redis==5.0.4
django-redis==5.4.0